// src/app/interfaces/valoracion.interface.ts
// Generated by https://quicktype.io

import { Cliente } from "./clientes.interface";
import { Variante } from "./facturas.interface";
import { Sucursal } from "./sucursal.interface";

export interface Valoracion {
  fechaDesde: string;
  fechaHasta: string;
  fechaCreacion: string;
  fechaModificacion: string;
  id: number | null;
  empresaId: number;
  sucursalId: number;
  listaPrecioId: number;
  varianteId: number;
  activo: boolean;
  cantDesde: number;
  cantHasta: number;
  valor: number;
  clienteId: number;
  registro: string;
  tipo: string;
  usuarioCreacion: null;
  usuarioModificacion: null;
  variante: Variante;
  listaPrecio: ListaPrecio;
  sucursal: Sucursal;
  cliente: Cliente;
  isSelected: boolean; // Asegúrate de inicializar isSelected
  isEdit: boolean; //
}

export interface ListaPrecio {
  descripcion: string;
}